# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse racing data correction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Race {
  id: string;
  race_date: string;
  venue: string;
  /** 中央競馬 | 地方競馬 | 海外競馬 */
  race_type: string;
  race_number: number;
  race_name: string;
  /** 芝 | ダート */
  surface_type: string;
  distance: number;
  /** @nullable */
  direction?: string | null;
  /** @nullable */
  weather?: string | null;
  /** @nullable */
  condition?: string | null;
  /** @nullable */
  start_time?: string | null;
  /** 未処理 | 補正中 | データ補正 | 補正完了 | 修正要求 | レビュー */
  status: string;
  /** @nullable */
  video_status?: string | null;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  analysis_status?: string | null;
  /** @nullable */
  assigned_user?: string | null;
  updated_at: string;
  created_at: string;
}

export interface UpdateRaceBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  assigned_user?: string | null;
  /** @nullable */
  video_status?: string | null;
}

export type RaceSummaryByVenueItem = {
  venue: string;
  count: number;
};

export interface RaceSummary {
  total: number;
  completed: number;
  in_progress: number;
  needs_correction: number;
  review: number;
  by_venue: RaceSummaryByVenueItem[];
}

export interface RaceEntry {
  id: string;
  race_id: string;
  horse_number: number;
  gate_number: number;
  horse_name: string;
  /** @nullable */
  jockey_name?: string | null;
  /** @nullable */
  trainer_name?: string | null;
  /** @nullable */
  last_3f?: number | null;
  /** @nullable */
  finish_time?: number | null;
  /** @nullable */
  finish_position?: number | null;
  /** @nullable */
  margin?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  lane?: string | null;
}

export interface PassingOrder {
  id: string;
  race_id: string;
  checkpoint: string;
  horse_number: number;
  horse_name: string;
  gate_number: number;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  lane?: string | null;
  /** @nullable */
  time_seconds?: number | null;
  /** @nullable */
  accuracy?: number | null;
  position: number;
  is_corrected: boolean;
  /** @nullable */
  original_position?: number | null;
}

export interface UpdatePassingOrderBody {
  /** @nullable */
  position?: number | null;
  /** @nullable */
  lane?: string | null;
  /** @nullable */
  time_seconds?: number | null;
}

export interface BatchJob {
  id: string;
  name: string;
  cron_expression: string;
  /** 有効 | 停止中 | 実行中 */
  status: string;
  is_enabled: boolean;
  /** @nullable */
  next_run_at?: string | null;
  created_at: string;
  updated_at: string;
}

export interface CreateBatchJobBody {
  name: string;
  cron_expression: string;
  is_enabled?: boolean;
}

export interface UpdateBatchJobBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  cron_expression?: string | null;
  /** @nullable */
  is_enabled?: boolean | null;
}

export interface Venue {
  id: string;
  name: string;
  race_type: string;
}

export type AnalysisParamsParams = { [key: string]: unknown };

export interface AnalysisParams {
  venue_id: string;
  venue_name: string;
  race_type: string;
  params: AnalysisParamsParams;
  updated_at?: string;
}

export type UpdateAnalysisParamsBodyParams = { [key: string]: unknown };

export interface UpdateAnalysisParamsBody {
  params: UpdateAnalysisParamsBodyParams;
}

export type GetRacesParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date?: string;
  /**
   * @nullable
   */
  venue?: string | null;
  /**
   * 中央競馬 | 地方競馬 | 海外競馬
   * @nullable
   */
  race_type?: string | null;
  /**
   * @nullable
   */
  video_status?: string | null;
  /**
   * @nullable
   */
  analysis_status?: string | null;
};

export type GetRaceSummaryParams = {
  date?: string;
};

export type GetPassingOrdersParams = {
  /**
   * Checkpoint name e.g. '200m'
   * @nullable
   */
  checkpoint?: string | null;
};