# lib/api-zod/src/generated/types/analysisParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse racing data correction API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisParamsParams } from "./analysisParamsParams";

export interface AnalysisParams {
  venue_id: string;
  venue_name: string;
  race_type: string;
  params: AnalysisParamsParams;
  updated_at?: string;
}