# lib/api-zod/src/generated/types/batchJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse racing data correction API
 * OpenAPI spec version: 0.1.0
 */

export interface BatchJob {
  id: string;
  name: string;
  cron_expression: string;
  /** 有効 | 停止中 | 実行中 */
  status: string;
  is_enabled: boolean;
  /** @nullable */
  next_run_at?: string | null;
  created_at: string;
  updated_at: string;
}