# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse racing data correction API
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: string;
  name: string;
  race_type: string;
}